requirements.txt
streamlit
llama-index-core
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-readers-file
python-dotenv
chromadb
pypdf
docx2txt
pysqlite3-binary